rename Filter package
