+++ src/main/java/Filter/EditFilterServlet.java
@@ -1,4 +1,4 @@
-package addFilter;
+package Filter;
 
 import java.io.IOException;
 import java.io.PrintWriter;
