+++ src/main/java/Filter/FilterNameCheckServlet.java
@@ -1,4 +1,4 @@
-package addFilter;
+package Filter;
 import com.example.*;
 import java.io.IOException;
 import java.io.PrintWriter;
